feat: re-structure project and correct the convention
