--- lib/openai/assistant/version.rb
@@ -1,7 +1,0 @@
-# frozen_string_literal: true
-
-module Openai
-  class Assistant < Openai::Base
-    VERSION = "0.7.0"
-  end
-end
