+++ lib/openai_assistant/base.rb
@@ -1,7 +1,5 @@
 # frozen_string_literal: true
 
-require_relative "../http/http"
-
 module Openai
   # Base class of openai
   class Base
@@ -10,7 +8,15 @@ module Openai
       # hard the host because if the official docs change the host, maybe it will change another
       # we need to update this gem for any change
       @openai_url = "https://api.openai.com/v1/assistants"
-      @http_client = Http::Client.new
+      @http_client = Openai::HTTPClient.new
+    end
+
+    def default_headers
+      {
+        "Authorization": "Bearer #{@openai_api_key}",
+        "OpenAI-Beta": "assistants=v1",
+        "Content-Type": "application/json"
+      }
     end
   end
 end
