+++ lib/openai_assistant/error_response.rb
