+++ lib/openai_assistant/version.rb
@@ -1,0 +1,5 @@
+# frozen_string_literal: true
+
+module Openai
+  VERSION = "0.7.0"
+end
