+++ spec/vcr_cassettes/create_assistant_valid.yml
@@ -9,7 +9,7 @@ http_interactions:
         write and run Ruby code to answer the question.","name":"assistant","tools":[{"type":"code_interpreter"}],"model":"gpt-3.5-turbo"}'
     headers:
       Authorization:
-      - Bearer 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
+      - Bearer <KP_TOKEN>
       Openai-Beta:
       - assistants=v1
       Content-Type:
@@ -26,7 +26,7 @@ http_interactions:
       message: OK
     headers:
       Date:
-      - Fri, 01 Dec 2023 09:53:55 GMT
+      - Tue, 05 Dec 2023 05:14:51 GMT
       Content-Type:
       - application/json
       Transfer-Encoding:
@@ -38,32 +38,32 @@ http_interactions:
       Openai-Organization:
       - user-lr17ucjtedcdrdzslupysvb8
       X-Request-Id:
-      - 40c91710b717f7b83e9da7c7f8a0bf4d
+      - e62cdd7d5ed51198469755f511bb0a86
       Openai-Processing-Ms:
-      - '59'
+      - '82'
       Strict-Transport-Security:
       - max-age=15724800; includeSubDomains
       Cf-Cache-Status:
       - DYNAMIC
       Set-Cookie:
-      - __cf_bm=22evnrM6cXecMD5tzmcGFjFZ_rYsRx1VIycBUVImjps-1701424435-0-ARgVNmhh/XrMug1C+e41xwujGzsf3aF1n4dVfqhWjQvQq8I+AAc0mepsfsGmopbHXOiVYNyd0ife/SfNuPsZnSk=;
-        path=/; expires=Fri, 01-Dec-23 10:23:55 GMT; domain=.api.openai.com; HttpOnly;
+      - __cf_bm=Uy6ViASfY2X711V7szK_DJxgwti2TgbSxhCZUNcm_0Q-1701753291-0-Acvq550DCWidEsljs0S3Pying1JVS0u1UuXP7Id7rIV8Wmxi+UbLfTUdxS20T21KkXN74FnFR6wqBi32HR7+XP8=;
+        path=/; expires=Tue, 05-Dec-23 05:44:51 GMT; domain=.api.openai.com; HttpOnly;
         Secure; SameSite=None
-      - _cfuvid=7y8qpDUve8HAEntKnur7UQRmBWZY_NF4Wq_mwcsTHXs-1701424435647-0-604800000;
+      - _cfuvid=F9n3AQwrWpLUWdEArqNV8Nmch48HWXqreo_7g5Y1Ofc-1701753291863-0-604800000;
         path=/; domain=.api.openai.com; HttpOnly; Secure; SameSite=None
       Server:
       - cloudflare
       Cf-Ray:
-      - 82ea72211f043e55-SIN
+      - 8309ced84e0d3e23-SIN
       Alt-Svc:
       - h3=":443"; ma=86400
     body:
       encoding: ASCII-8BIT
       string: |-
         {
-          "id": "asst_ORdKV3nCiN8WwtpfOqvpT3jn",
+          "id": "asst_R4kiJBbQuFQ0QBa9pDJXBane",
           "object": "assistant",
-          "created_at": 1701424435,
+          "created_at": 1701753291,
           "name": "assistant",
           "description": null,
           "model": "gpt-3.5-turbo",
@@ -76,5 +76,5 @@ http_interactions:
           "file_ids": [],
           "metadata": {}
         }
-  recorded_at: Fri, 01 Dec 2023 09:53:55 GMT
+  recorded_at: Tue, 05 Dec 2023 05:14:51 GMT
 recorded_with: VCR 6.2.0
